@@ -32,32 +32,32 @@
         "build:lambda": "netlify-lambda build src/functions"
     },
     "dependencies": {
-        "@fortawesome/fontawesome-svg-core": "^1.2.12",
-        "@fortawesome/free-brands-svg-icons": "^5.6.3",
-        "@fortawesome/free-solid-svg-icons": "^5.6.3",
-        "@fortawesome/react-fontawesome": "^0.1.3",
+        "@fortawesome/fontawesome-svg-core": "^1.2.15",
+        "@fortawesome/free-brands-svg-icons": "^5.7.2",
+        "@fortawesome/free-solid-svg-icons": "^5.7.2",
+        "@fortawesome/react-fontawesome": "^0.1.4",
         "babel-plugin-styled-components": "^1.10.0",
         "dotenv": "^6.2.0",
-        "gatsby": "^2.0.76",
-        "gatsby-plugin-google-analytics": "^2.0.8",
+        "gatsby": "^2.1.4",
+        "gatsby-plugin-google-analytics": "^2.0.14",
         "gatsby-plugin-google-fonts": "0.0.4",
-        "gatsby-plugin-react-helmet": "^3.0.4",
-        "gatsby-plugin-sass": "^2.0.7",
-        "gatsby-plugin-styled-components": "^3.0.4",
-        "gatsby-source-contentful": "^2.0.21",
-        "gatsby-source-shopify": "^2.0.11",
-        "markdown-to-jsx": "^6.9.0",
-        "netlify-lambda": "^1.1.1",
+        "gatsby-plugin-react-helmet": "^3.0.6",
+        "gatsby-plugin-sass": "^2.0.10",
+        "gatsby-plugin-styled-components": "^3.0.5",
+        "gatsby-source-contentful": "^2.0.29",
+        "gatsby-source-shopify": "^2.0.13",
+        "markdown-to-jsx": "^6.9.1",
+        "netlify-lambda": "^1.4.2",
         "node-sass": "^4.11.0",
-        "npm": "^6.5.0",
-        "react": "^16.6.3",
-        "react-dom": "^16.6.3",
+        "npm": "^6.8.0",
+        "react": "^16.8.2",
+        "react-dom": "^16.8.2",
         "react-helmet": "^5.2.0",
         "shopify-buy": "^2.0.1",
         "styled-components": "^4.1.3"
     },
     "devDependencies": {
-        "eslint": "^5.11.0",
-        "eslint-plugin-react": "^7.11.1"
+        "eslint": "^5.14.1",
+        "eslint-plugin-react": "^7.12.4"
     }
 }
